+++ execution_engine_testing/cargo_casper/CHANGELOG.md
@@ -16,6 +16,9 @@ All notable changes to this project will be documented in this file.  The format
 ### Changed
 * Support building and testing using stable Rust.
 
+### Removed
+* Remove support for generating an ERC-20 contract.
+
 
 
 ## [1.3.3] - 2021-09-13